refactor: 🚚 Déplaçage du typage du DOM dans le dossier types
